Show centered logo header on all public pages
