+++ app/src/app/annuler/[token]/page.tsx
@@ -1,4 +1,5 @@
 import { notFound } from "next/navigation";
+import SiteHeader from "@/components/SiteHeader";
 import { prisma } from "@/lib/prisma";
 import { formatDateFr, formatTimeFr } from "@/lib/dates";
 import CancelActions from "./CancelActions";
@@ -10,8 +11,9 @@ export default async function CancelPage({ params }: { params: { token: string }
   if (!booking) notFound();
 
   return (
-    <main className="mx-auto flex min-h-screen max-w-lg flex-col justify-center px-4 py-10">
-      <h1 className="text-2xl font-extrabold">Annuler ou reporter</h1>
+    <main className="mx-auto flex min-h-screen max-w-lg flex-col justify-center px-4 py-6">
+      <SiteHeader />
+      <h1 className="mt-4 text-2xl font-extrabold">Annuler ou reporter</h1>
       <div className="card mt-5 space-y-1">
         <p className="font-semibold">
           📅 {formatDateFr(booking.startAt)} à {formatTimeFr(booking.startAt)}
